pyproject.toml
[tool.poetry]
name = "harmonica"
version = "0.1.0"
description = ""
authors = ["vaclav-v"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
python-telegram-bot = {git = "https://github.com/python-telegram-bot/python-telegram-bot.git", rev = "master"}
pillow = "^9.5.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
